chore: separate types from Music Together
