+++ src/plugins/music-together/index.ts
@@ -6,7 +6,7 @@ import { t } from '@/i18n';
 import { createPlugin } from '@/utils';
 import promptOptions from '@/providers/prompt-options';
 
-import { type AppElement, getDefaultProfile, type Permission, type Profile, type VideoData } from './types';
+import { getDefaultProfile, type Permission, type Profile, type VideoData } from './types';
 import { Queue } from './queue';
 import { Connection, type ConnectionEventUnion } from './connection';
 import { createHostPopup } from './ui/host';
@@ -19,6 +19,7 @@ import style from './style.css?inline';
 import type { YoutubePlayer } from '@/types/youtube-player';
 import type { RendererContext } from '@/types/contexts';
 import type { VideoDataChanged } from '@/types/video-data-changed';
+import type { AppElement } from '@/types/queue';
 
 type RawAccountData = {
   accountName: {
